Implement premium calculation Add happy path unit tests

+++ src/main/java/com/proofit/policy/premium/dto/Policy.java
@@ -9,6 +9,6 @@ public class Policy {
 
     private String policyNumber;
     private List<PolicyObject> policyObjects;
-    private long premium;
+    private double premium;
 
 }
